1、添加待处理交易列表及详情接口

+++ common/src/main/java/com/platon/browser/common/dto/block/BlockList.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class BlockList {
-    private int height;
+    private long height;
     private long timestamp;
     private int transaction;
     private int size;
