+++ common/src/main/java/com/platon/browser/common/dto/transaction/PendingTxList.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class PendingTxList {
     private String txHash;
-    private int dwellTime;
+    private long dwellTime;
     private int energonLimit;
     private int energonPrice;
     private String from;
